https://github.com/chaotic-anger/pet-laravel/issues/9 вынос в Feature-каталог
