+++ app/Api/Enums/VoteDirection.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 
-namespace App\Enums;
+namespace App\Api\Enums;
 
 
 enum VoteDirection: string
