+++ app/Api/Events/CommentVoteCreated.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Events;
+namespace App\Api\Events;
 
-use App\Models\CommentVote;
+use App\Api\Models\CommentVote;
 use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
