+++ app/Api/Factory/CommentFactory.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Database\Factories;
+namespace App\Api\Factory;
 
-use App\Models\Comment;
+use App\Api\Models\Comment;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Comment>
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Api\Models\Comment>
  */
 class CommentFactory extends Factory
 {
