+++ app/Api/Http/Controllers/PostController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Controllers\Api\Api;
+namespace App\Api\Http\Controllers;
 
+use App\Api\Http\Resources\PostResource;
+use App\Api\Models\Post;
 use App\Http\Controllers\Controller;
-use App\Http\Resources\PostResource;
-use App\Models\Post;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
