+++ app/Providers/AppServiceProvider.php
@@ -2,7 +2,12 @@
 
 namespace App\Providers;
 
-use App\Listeners\CommentVoteSubscriber;
+use App\Api\Events\Listeners\CommentVoteSubscriber;
+use App\Api\Models\Comment;
+use App\Api\Models\Post;
+use App\Api\Policies\CommentPolicy;
+use App\Api\Policies\PostPolicy;
+use Gate;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -13,6 +18,8 @@ class AppServiceProvider extends ServiceProvider
 
     public function boot(): void
     {
+        Gate::policy(Post::class, PostPolicy::class);
+        Gate::policy(Comment::class, CommentPolicy::class);
         $this->app['events']->subscribe(CommentVoteSubscriber::class);
     }
 }
