+++ bootstrap/app.php
@@ -28,5 +28,5 @@ return Application::configure(basePath: dirname(__DIR__))
         ]);
     })
     ->withExceptions(function (Exceptions $exceptions) {
-        $exceptions->render(new \App\Exceptions\ApiErrorHandler());
+        $exceptions->render(new \App\Api\Exceptions\ApiErrorHandler());
     })->create();
